Megatron refactor: direct imports, move utils, rename configs
- Replace create_training_job factory with direct MegatronTrainingJob in megatron_single.py and megatron_distributed.py (primus not yet implemented)
- Move model_registry.py and training_config_loader.py to megatron/utils/ and update imports in megatron_lib.py and conftest.py
- Rename llama3_3_70b config and threshold files to mi325x_megatron_llama-3.3-70b_single naming convention
