+++ cvs/tests/training/megatron/conftest.py
@@ -13,7 +13,7 @@ import pytest
 from cvs.core.orchestrators.factory import OrchestratorConfig, OrchestratorFactory
 from cvs.lib import globals
 from cvs.lib.utils_lib import resolve_cluster_config_placeholders
-from cvs.lib.training.megatron.training_config_loader import load_training_variant
+from cvs.lib.training.megatron.utils.training_config_loader import load_training_variant
 
 log = globals.log
